# Use method BindMessageTemplate to improve performance and memory usage