@nblumhardt 2021-05-18
> apologies in advance if the turnaround is not snappy

That may have been the understatement of 2020! :-D ... Thanks again for this, Maxim.

I finally did have time to think it through; my main concern with using `BindMessageTemplate()` here is that it makes the enumeration order of the event properties important. It's not clear that the order of properties in the enumerated structure will always match the order of property names in the template holes. Some `ILogger<T>` optimization techniques hand-code the enumeration of properties in the state object, so this could be broken even if we don't have a failing case in front of us.

I don't think we should proceed in this direction, but thanks again, and please let me know if it sounds like I've missed something.

Best regards,
Nick